moved info mode code from standalone to buildinfo package
